+++ src/renderer/contexts/CollapsedNodeContext.tsx
@@ -1,8 +1,8 @@
 import { Box } from '@chakra-ui/react';
 import React, { memo } from 'react';
-import { createContext, useContext } from 'use-context-selector';
+import { createContext } from 'use-context-selector';
 
-const IsCollapsedContext = createContext<boolean>(false);
+export const IsCollapsedContext = createContext<boolean>(false);
 
 export const CollapsedNode = memo(({ children }: React.PropsWithChildren<unknown>) => {
     return (
@@ -14,7 +14,3 @@ export const CollapsedNode = memo(({ children }: React.PropsWithChildren<unknown
         </Box>
     );
 });
-
-export const useIsCollapsedNode = (): boolean => {
-    return useContext(IsCollapsedContext);
-};
