+++ src/renderer/hooks/useIsCollapsedNode.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'use-context-selector';
+import { IsCollapsedContext } from '../contexts/CollapsedNodeContext';
+
+export const useIsCollapsedNode = (): boolean => {
+    return useContext(IsCollapsedContext);
+};
